+++ libraries/base/changelog.md
@@ -22,6 +22,9 @@
 
   * Re-export `Data.Word.Word` from `Prelude`
 
+  * Add `countLeadingZeros` and `countTrailingZeros` methods to
+    `Data.Bits.FiniteBits` class
+
 ## 4.7.0.1  *Jul 2014*
 
   * Bundled with GHC 7.8.3